Group Split
This code is after spliting with Vitoria.
The initial commit of this project has been taken from another repository where there was a collaboration work between me (Joshua Reyes) and Vitoria Oliveira, the spliting decision was made to diffculties to work in team only in a remote way and having different working time tables, we found it highly inneficient in terms of time.

The code done in teams was to generate all the animals and staff at the begining;
Animals and Staff categories are divided in different packages;
Each category has a super class and one subclass. The subclass holds inner classes for each kind of animal and staff respectively;
There is a class in each package to generate all the infomation of each instance of animal or staff;
In the default package there is a class called helper which ill be used to generate all the information that is going to be used for the user to interact with the program.

Some other changes:
-Added a new attribute for the Admin staff which will save the activity their are preforming at the moment.

-Added a new class that generates an activiy for the admin staff

-Modified the menu